+++ tests/providers/amazon/aws/hooks/test_glue_catalog.py
@@ -21,7 +21,7 @@ import unittest
 import boto3
 import mock
 
-from airflow.contrib.hooks.aws_glue_catalog_hook import AwsGlueCatalogHook
+from airflow.providers.amazon.aws.hooks.glue_catalog import AwsGlueCatalogHook
 
 try:
     from moto import mock_glue
